+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## v1.2.0
+
+- Adding `overlay` functionality and documentation
+- Enable printing of CSS background colors/image
+- Dependencies upgrade
+
 ## v1.1.2
 
-- Fix [Issue 3](https://github.com/motla/vue-document-editor/issues/2)
+- Fix [Issue 3](https://github.com/motla/vue-document-editor/issues/3)
   - Renaming the static library "VueDocumentEditor"
   - Adding UMD example to README.md